rework lcw vc population
